separated color functions into another file. also automatically calculate contrasting text color.
